Refactor directory structure to suit webpack
Move JavaScript and CSS files to `src/`. Delete excess HTML files and
create template in `views/`. Create npm script to enable webpack's watch
option. Note that this only works for files in the Linux filesystem on
WSL-2.

This commit is not suitable for production. No bugs have been fixed in
the scripts or stylesheets as these will be fixed by the merge commit
with `misc-pages`.