Add functionality and improve account page.
Add header and footer components to the account page.
Add id="loginlink" to anchor of header login element in common-sections.jsx.
Using javascript, replace innerHTML and href of "loginlink".
Change innerHTML to "Log Out" and href to "/logout".
Remove the sessionStorage containing the user's name when the log out button is clicked.
Use the useEffect hook to run the javascript after the html is loaded.